Add own packages to ignorePackages, and move some tests into a different package so that they still work. May revert if doesn't help.
