+++ src/test/java/nz/ac/wgtn/veracity/provenance/injector/testinstrumentation/InvocationCreationTestCase.java
@@ -1,5 +1,7 @@
-package nz.ac.wgtn.veracity.provenance.injector.instrumentation;
+package nz.ac.wgtn.veracity.provenance.injector.testinstrumentation;
 
+import nz.ac.wgtn.veracity.provenance.injector.instrumentation.AssociationCache;
+import nz.ac.wgtn.veracity.provenance.injector.instrumentation.AssociationCacheRegistry;
 import nz.ac.wgtn.veracity.provenance.injector.sampleclasses.SomeClass;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
